Move the common functions to helpers
